简化代码

+++ src/main/java/com/example/entity/Employee.java
@@ -28,7 +28,7 @@ public class Employee {
     @Column
     private String employeeSex;
     @Column
-    private Integer employeeAge;
+    private Byte employeeAge;
     @Column
     private String employeeIdCard;
     @Column
